upgrade to new schema bundle format
